requirements.txt
google-generativeai
langgraph
httpx
langchain-google-genai
einops
timm
huggingface-hub
wheel
cython
flash_attn
transformers
pandas
numpy
matplotlib
seaborn
pillow
requests
streamlit
ipython
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118